Add files via upload
This is placement cell website which basically deals with the placement regarding the students of the college. The main aim of developing this system was to digitally enterprise the work done by the placement cell, its staff and the volunteers. The photos of the system are not being attached as the file was a bit bigger in size, so anyone who wishes to downloads this project can make or implement his/her own changes.